Add new authentication method notice
This is notice to remind users that after switching to a new authentication method
they have to import new auth method users and set admin role to at least one of them.
Otherwise user will loose administrative access to the web interface.

The idea to add this notice was reported by lucasljorge on Bacularis User Group:

https://group.bacularis.app/d/126-issue-with-authentication